FUncionamiento para recuperar una comunidad por ID

+++ src/main/java/com/comunidad/ad2/comunidad/repository/ComunityRepository.java
@@ -15,10 +15,11 @@ import org.springframework.test.context.jdbc.Sql;
  *
  * @author jesfrin
  */
-public interface ComunityRepository extends JpaRepository<Comunity, Long> {
+public interface ComunityRepository extends JpaRepository<Comunity, Integer> {
 
     //@Query("SELECT c FROM Comunity com join com.user WHERE com.id=:id")u.registroAcademico=?1 
     @Query("SELECT comunity from Comunity comunity where comunity.user.registroAcademico=?1")
     public Iterable<Comunity> findByRegistroAcademico(String registroAcademico);
 
+     
 }
